+++ doc/examples/myRModel.R
@@ -29,6 +29,8 @@ myModel <- function(pars){
 mySummaryStatistics <- function(modelResult){
   list(x=modelResult[1],
        y=modelResult[2],
+       mtcars=mtcars,  # Can also pass data frames
+       iris=iris,
        arbitraryKey="Some random text")
 }
 
@@ -49,7 +51,7 @@ myDistance <- function(sumStatSample, sumStatData){
 # in a variable.
 # This is passed by pyABC as to myDistance
 # as the sumStatData argument
-mySumStatData <- list(x=4, y=8)
+mySumStatData <- list(x=4, y=8, mtcars=mtcars, iris=iris)
 
 # The functions for the model, the summary
 # statistics and distance